// types/user.ts
export interface User {
  id: string
  name: string
  email: string
  password: string
  age: number
  gender: string
  height: number
  weight: number
  fitnessGoal: string
  fitnessLevel: string
  workoutLocation: string
  dietaryPreference: string
  medicalHistory: string
  stressLevel: string
  plan: any
  createdAt: string
}